Changing BigDecimal to int since amount is in cents.

+++ ap-java-sdk/src/main/java/com/alternativepayments/models/plan/Plan.java
@@ -1,7 +1,5 @@
 package com.alternativepayments.models.plan;
 
-import java.math.BigDecimal;
-
 import org.joda.time.DateTime;
 
 import com.alternativepayments.models.BaseModel;
@@ -21,7 +19,7 @@ public class Plan extends BaseModel {
 
     private final int interval;
     private final Period period;
-    private final BigDecimal amount;
+    private final int amount;
     private final String currency;
     private final String name;
     private final String description;
@@ -48,7 +46,7 @@ public class Plan extends BaseModel {
             @JsonProperty("updated") final DateTime updated,
             @JsonProperty("interval") final int interval,
             @JsonProperty("period") final Period period,
-            @JsonProperty("amount") final BigDecimal amount,
+            @JsonProperty("amount") final int amount,
             @JsonProperty("currency") final String currency,
             @JsonProperty("name") final String name,
             @JsonProperty("description") final String description) {
@@ -68,7 +66,7 @@ public class Plan extends BaseModel {
     public static class Builder {
         // Required parameters
         private final String name;
-        private final BigDecimal amount;
+        private final int amount;
         private final String currency;
         private final int interval;
         private final Period period;
@@ -85,7 +83,7 @@ public class Plan extends BaseModel {
          * @param interval interval for the plan
          * @param period period for the plan
          */
-        public Builder(final String name, final BigDecimal amount, final String currency, final int interval,
+        public Builder(final String name, final int amount, final String currency, final int interval,
                 final Period period) {
             this.name = name;
             this.amount = amount;
@@ -147,7 +145,7 @@ public class Plan extends BaseModel {
     /**
      * @return the amount
      */
-    public BigDecimal getAmount() {
+    public int getAmount() {
         return amount;
     }
